refactor: Changed middleware implementation with FormRequest Validation, simpler implementation of json response by creating a new method, optimize code, adjust kernel, route, and related
