+++ app/Http/Middleware/ForceJsonResponse.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+
+class ForceJsonResponse
+{
+  public function handle(Request $request, Closure $next)
+  {
+    $request->headers->set('Accept', 'application/json');
+
+    return $next($request);
+  }
+}
