+++ app/Services/VehicleService.php
@@ -3,7 +3,7 @@
 namespace App\Services;
 
 use App\Helpers\FeeHelper;
-use App\Services\FeeService;
+use App\Enums\FeeType;
 use App\Http\Requests\CalculateFeeRequest;
 
 class VehicleService
@@ -20,12 +20,39 @@ class VehicleService
 
     $totalCost = $vehicleBasePrice + $basicFee + $specialFee + $associationFee + $storageFee;
 
-    $fees = FeeService::generateFees($basicFee, $specialFee, $associationFee, $storageFee);
+    $fees = self::generateFees($basicFee, $specialFee, $associationFee, $storageFee);
 
     return response()->json([
       'fees' => $fees,
       'vehicle_price' => $vehicleBasePrice,
       'total_price' => $totalCost,
     ]);
+  }
+
+  private static function generateFees($basicFee, $specialFee, $associationFee, $storageFee)
+  {
+    $basicFeeType = FeeType::Basic->value;
+    $specialFeeType = FeeType::Special->value;
+    $associationFeeType = FeeType::Association->value;
+    $storageFeeType = FeeType::Storage->value;
+
+    return [
+      [
+        "type" => $basicFeeType,
+        "amount" => $basicFee,
+      ],
+      [
+        "type" => $specialFeeType,
+        "amount" => $specialFee,
+      ],
+      [
+        "type" => $associationFeeType,
+        "amount" => $associationFee,
+      ],
+      [
+        "type" => $storageFeeType,
+        "amount" => $storageFee,
+      ],
+    ];
   }
 }
